requirements.txt
allennlp==1.0
transformers==2.11
wandb
pyyaml
pandas
tabulate
lxml
-e git+git@github.com:pasinit/nlp_tools.git#egg=nlp_tools